update: number of images is no longer constrained to be width x height
Long overdue update so that grid allows blank spaces which
are intelligently assigned away from the center. Updated to
work with aspect-ratio, etc. but the --show-links option
is almost certainly broken. Currently, blank spaces are
filled with a dummy image which is dymanically created -
the level of gray can be specified with the --fill-shade
option. There is also a --drop-to-fit which keeps the
old behavior of dropping images to force fitting to
a rectangle.